backend/requirements.txt
# MemeVault Backend - BLUEPRINT uyumlu
# Python 3.11+

# Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Veritabanı
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pgvector>=0.2.4

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Celery & Redis
celery[redis]>=5.3.0
redis>=5.0.0

# Object Storage (S3 uyumlu)
boto3>=1.34.0

# Ingestion
yt-dlp>=2024.1.0

# AI / Embedding (MVP: doğrudan API)
openai>=1.12.0
httpx>=0.26.0

# Config & env
pydantic-settings>=2.1.0

# Yardımcı
python-multipart>=0.0.6
uuid7>=0.1.0
